requirements.txt
Django==1.11.10
django-braces==1.12.0
django-cors-headers==2.1.0
djangorestframework==3.7.7
mixin==1.1
Pillow==5.0.0
pytz==2017.3
